wine-pure: Drop patch to fix GL performance in WoW64 when using Zink
This does solve the performance issue, but breaks winewayland in a
number of games
(https://gitlab.winehq.org/wine/wine/-/merge_requests/6663#note_86455).

Also this solution probably won't be merged into upstream, and the only
driver that supports it is Zink, which implies that your GPU should
already support Vulkan and that doesn't help people who don't have
support for it. The best thing to do right now is to use Vulkan or
DX9-11 (to combine it with DXVK) translation wrappers like dgVoodoo2 or
cnc-ddraw.